backend/requirements.txt
django
django-cors-headers
djangorestframework
channels
Pillow
